Gradle Build: Build and store the offline repo.
The gmaven zip contains the artifacts we publish to Google's maven
repository. This includes the Android Gradle Plugin and various other
libraries, but does not inlcude the prebuilt transitive dependencies.

The offline repo contains the artifacts needed to run the Android Gradle
Plugin, and so it additionally contains prebuilt transitive
dependencies, but does not contain extra libraries built from
studio-master-dev that are not needed by the Android Gradle Plugin.

This change builds the offline repo and points the Gradle Build to use
the dist directory so that it appears in the artifacts on the Android
Build server.

Bug: N/A
Test: Manually verified artifact is present.
Change-Id: If58362506a44cb8df3b3e78ca171ff77ca77c222