Renames BUILD to BUILD.bazel
For some reason buildSrc/base is added as a resource
of android.sdk.common into the com/android directory.

This conflicts with a "build" directory at runtime that comes
from gradle in case insensitive file systems.

Unclear why we pull this in as a resource, but this rename
fixes this issue.

Bug: 156497102
Test: existing
Change-Id: I3761933731d314c638c4783b950c4dba93fd99e2