Update "requires JDK 11" message with how to change it
I tried opening the tools project in the IDE and it told
me I needed to use Java 11 instead of Java 8. I spent
an embarassing amount of time trying to track this down,
updating environment variables, trying to launch the
IDE from outside JB Toolbox, verifying envvars from
within the embedded terminal emulator, etc. Turns out
this is simply a sync setting in the IDE. This CL changes
the error message to hint about that.

Test: N/A
Bug: None
Change-Id: I1e9022235d27af317887c99b96d13ff671643256