Rename the :base:lint project to :base:lint-cli
For historical reasons the CLI library of lint was just
named "lint" (to match the binary), but the corresponding
Gradle artifact and source tree location included "cli".
This CL renames the Gradle project to :base:lint-cli to
reflect this (and adds a new :base:lint project which just
contains root tasks for all the lint projects; currently
formatting and possibly in the future metalava API tracking.

Test: N/A
Bug: N/A
Change-Id: I273c032c6cc9305ba8ed0976f1193e0e49193e22