Add the studio-checks lint project
We have a number of lint checks intended to be run on our own codebase.
These are developed in the lint/studio-checks project, but that project
was only built by bazel. This CL also registers this project with Gradle
such that it's part of the tools/ project, and participates in find
usages, refactoring, you can run its unit tests, etc. The project does
not apply any publishing plugins; it's not meant to be distributed.

Bug: None
Test: N/A
Change-Id: Ifa82e4fb2b26967db063444ecd7395f70b975929