Disable Kotlin sources task
As it overwrites the java library sources jar, which has the form we
want for now.

This fix can be manually verified by running

    gradlew -Prelease=true \
        :b:b-s:g-c:sourcesJar :b:b-s:g-c:kotlinSourcesJar \
        :pAGL

which means the tasks will run in the order that leads to this issue and
verifying the content of the sources jar in out/repo e.g.

    jar tf $SRC/out/repo/com/android/tools/build/gradle/*/gradle-*-sources.jar

Bug: 276283430
Test: See tools/base
Change-Id: I674860b61ef0ae950c0290746147525607e97814