AGP build: Move "out/build" directory to inside "tools" directory
Before this change, the structure of AGP repo is:
  - studio-main
      + out
         + build // ORIGIN
         + repo
      + tools
      + etc.

After this change, "out/build" will be moved to "tools/build":
  - studio-main
      + out
         + repo
      + tools
         + build // TARGET
      + etc.

This is to work around https://youtrack.jetbrains.com/issue/KT-61852
which does not allow source sets (in this case, generated protos in the
build directory) to be located outside the project directory.

This bug breaks incremental compilation of AGP (incremental compilation
always fails and falls back to non-incremental compilation).

This workaround allows us to restore incremental Kotlin compilation when
building AGP.

Bug: 301913541
Test: `./gradlew publishLocal` produces outputs in "tools/build"
Change-Id: I316992921e25db154099f7a73bb1c046bf3757ee