feat: Add support for symbol extraction and upload
This commit introduces a new target that extracts symbols from the build
files, enabling the use of the crashpad crash reporting system for
enhanced debugging capabilities.

The build system now creates a zip file containing the symbols:

* **Windows:** sdk-repo-{platform}-{product_name}-pdbs-{build_id}.zip
* **Other platforms:** sdk-repo-{platform}-{product_name}-breakpad-symbols-{build_id}.zip

These symbols are automatically uploaded to a symbol server for both
presubmit and release builds, using the following endpoints:

* **Presubmit:** https://staging-crashsymbolcollector-pa.googleapis.com
* **Release:** https://prod-crashsymbolcollector-pa.googleapis.com

To authenticate with the symbol server, the build system requires an API
key, which can be provided either through the
`EMULATOR_SYMBOL_SERVER_KEY` environment variable or by placing it in
the `~/.emulator_symbol_server_key` file.

Bug: 331193952
Change-Id: I70d0ec94983046edb8ae635f32e23fb87d2c6f0d