Migrate off deprecated --api-version-names options
Previously, `--api-version-signature-files` and `--api-version-names`
were required to be provided together as each item in the latter
provided the version name for the corresponding file in the format.
That complicated both this code and Metalava. This change switches to
use the `--api-version-signature-pattern` which gives Metalava the
information it needs to extract the version from the files, removing
the need for `--api-version-names` altogether.

The `--api-version-names` was also used to provide the current version,
i.e. the version of the sources. That is now done using
`--current-version`.

Verified as follows:
```
    # Before applying this change.
    cd ../base/build-system
    rm ../../../out/build/base/build-system/gradle-settings-api/build/metalava/current/apiLevels.json
    ../../gradlew generateApi
    mkdir before
    mv ../../../out/build/base/build-system/gradle-settings-api/build/metalava/current/apiLevels.json before

    # After applying this change.
    ../../gradlew generateApi
    mkdir after
    cp ../../../out/build/base/build-system/gradle-settings-api/build/metalava/current/apiLevels.json after

    # Verified that the following produced no output.
    diff -r before after
```

Bug: 388441272
Test: See above
Change-Id: Ie61788afcf6f8b9fc74303d7ac31f81bad573678